Moved FunctorTask to its own file
Moved the helper class FunctorTask to its own file in order to use it more broadly.
